auth login controller

+++ API_wee/Controllers/AuthController.cs
@@ -166,7 +166,7 @@ namespace API_wee.Controllers
             public string Username { get; set; } = null!;
             public string LastName { get; set; } = null!;
             public string Password { get; set; } = null!;
-            public int RoleId { get; set; }
+            public short RoleId { get; set; }
         }
     }
 }
