+++ API_wee/Services/IUserService.cs
@@ -6,7 +6,7 @@ namespace API_wee.Services
     public interface IUserService
     {
         Task<ApplicationUser?> AuthenticateAsync(string username, string password);
-        Task<ApplicationUser> CreateUserAsync(string username, string lastname, string password, int roleId);
+        Task<ApplicationUser> CreateUserAsync(string username, string lastname, string password, short roleId);
         Task<ApplicationUser?> GetByIdAsync(int id);
     }
 }
